api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.2
opencv-python-headless==4.8.1.78
psutil==5.9.6
